Backend/requirements.txt
Flask==3.0.0
Flask-CORS==4.0.0
Flask-JWT-Extended==4.6.0
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.5
python-dotenv==1.0.0
bcrypt==4.1.2
PyJWT==2.8.0
requests==2.31.0
python-dateutil==2.8.2
Werkzeug==3.0.1
gunicorn==21.2.0
pytest==7.4.3
pytest-flask==1.3.0
google-generativeai==0.3.0
python-jose==3.3.0
cryptography==42.0.5
reportlab==4.0.8 